refactor/add customization tests
