+++ tests/indicators/_common/Quotes/Quote.Converters.Tests.cs
@@ -1,11 +1,14 @@
 using System.Collections.ObjectModel;
 
+// quote list converters
+
 namespace Utilities;
 
+[TestClass]
 public partial class Quotes : TestBase
 {
     [TestMethod]
-    public void QuoteToSortedCollection()
+    public void ToSortedCollection()
     {
         IReadOnlyList<Quote> quotes = Data.GetMismatch();
 
@@ -28,7 +31,7 @@ public partial class Quotes : TestBase
     }
 
     [TestMethod]
-    public void QuoteToSortedList()
+    public void ToSortedList()
     {
         IReadOnlyList<Quote> quotes = Data.GetMismatch();
