+++ tests/indicators/_common/Quotes/Quote.Exceptions.Tests.cs
@@ -1,20 +1,21 @@
 namespace Utilities;
 
+// invalid quotes exceptions
+
 public partial class Quotes : TestBase
 {
-    // invalid quotes exceptions
     [TestMethod]
     [ExpectedException(typeof(InvalidQuotesException), "Bad quotes without message.")]
-    public void ThrowInvalidQuotesException()
+    public void InvalidQuotesExceptionThrow()
         => throw new InvalidQuotesException();
 
     [TestMethod]
     [ExpectedException(typeof(InvalidQuotesException), "Bad quotes with message.")]
-    public void ThrowInvalidQuotesExWithMessage()
+    public void InvalidQuotesExceptionThrowWithMessage()
         => throw new InvalidQuotesException("This is a quotes exception.");
 
     [TestMethod]
     [ExpectedException(typeof(InvalidQuotesException), "Bad quotes with inner Exception.")]
-    public void ThrowInvalidQuotesExWithInner()
+    public void InvalidQuotesExceptionThrowWithInner()
         => throw new InvalidQuotesException("This has an inner Exception.", new ArgumentException());
 }
