+++ tests/indicators/_common/Quotes/Quote.StreamHub.Tests.cs
@@ -1,5 +1,8 @@
 namespace StreamHub;
 
+// QUOTEHUB
+
+[TestClass]
 public class QuoteHub : StreamHubTestBase, ITestChainProvider
 {
     [TestMethod]
