--- tests/indicators/_common/Quotes/Quote._base.StaticSeries.Tests.cs
@@ -1,6 +1,0 @@
-namespace Utilities;
-
-[TestClass]
-public partial class Quotes : TestBase
-{
-}
